moved Config to configuration package
