+++ app/src/main/java/org/stepik/android/adaptive/App.kt
@@ -2,6 +2,7 @@ package org.stepik.android.adaptive
 
 import android.support.multidex.MultiDexApplication
 import org.solovyev.android.checkout.Billing
+import org.stepik.android.adaptive.configuration.Config
 
 class App : MultiDexApplication() {
     val billing by lazy {
